requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.9
streamlit==1.28.1

# Data processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.4

# Time series and ML
statsmodels==0.14.0
xgboost==2.0.2
joblib==1.3.2

# IoT and messaging
paho-mqtt==1.6.1
requests==2.31.0

# Database
timescaledb==0.0.1
alembic==1.12.1

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0



python-jose[cryptography]
email-validator